Updated test to use sys_bigint
