+++ zkvm/entrypoint/src/syscalls/uint256_mul.rs
@@ -20,32 +20,3 @@ pub extern "C" fn syscall_uint256_mul(x: *mut u32, y: *const u32) {
     #[cfg(not(target_os = "zkvm"))]
     unreachable!()
 }
-
-const BIGINT_WIDTH_WORDS: usize = 8;
-
-#[allow(unused_variables)]
-#[no_mangle]
-pub extern "C" fn sys_bigint(
-    result: *mut [u32; BIGINT_WIDTH_WORDS],
-    op: u32,
-    x: *const [u32; BIGINT_WIDTH_WORDS],
-    y: *const [u32; BIGINT_WIDTH_WORDS],
-    modulus: *const [u32; BIGINT_WIDTH_WORDS],
-) {
-    let mut concat_y_modulus = core::mem::MaybeUninit::<[u32; BIGINT_WIDTH_WORDS]>::uninit();
-    unsafe {
-        let result_ptr = result as *mut u32;
-        let x_ptr = x as *const u32;
-        let y_ptr = y as *const u32;
-        let concat_ptr = concat_y_modulus.as_mut_ptr() as *mut u32;
-
-        core::ptr::copy_nonoverlapping(y_ptr, concat_ptr, BIGINT_WIDTH_WORDS);
-        core::ptr::copy_nonoverlapping(
-            modulus as *const u32,
-            concat_ptr.add(BIGINT_WIDTH_WORDS),
-            BIGINT_WIDTH_WORDS,
-        );
-        core::ptr::copy_nonoverlapping(x as *const u32, result_ptr, BIGINT_WIDTH_WORDS);
-        syscall_uint256_mul(result_ptr, concat_ptr);
-    }
-}
